Add empty spot shows ghost course when dragging course
