+++ app/dashboard/(schedule)/(block)/StandardBlock.tsx
@@ -1,4 +1,4 @@
-import { Draggable } from "@/components/CourseCard/Draggable";
+import { Draggable } from "@/components/DndProvider/Draggable";
 import { Droppable } from "@/components/Droppable";
 import CourseCard from "@/components/CourseCard";
 import { SearchIcon } from "lucide-react";
@@ -16,7 +16,7 @@ const StandardBlock: FC<BlockViewProps> = ({
     return (
       <Droppable data={data} id={blockId}>
         <Draggable id={courseSlot.code} data={courseSlot}>
-          <CourseCard course={courseSlot} dropped={true} />
+          <CourseCard variant="dropped" course={courseSlot} />
         </Draggable>
       </Droppable>
     );
