+++ components/CourseModal/DialogFooterWithDetails.tsx
