+++ components/CourseModal/DialogGeneralTab.tsx
@@ -1,7 +1,7 @@
 import DialogTimeChart from "./DialogTimeChart";
 import { Label } from "@/components/ui/label";
 import { Course } from "@/app/dashboard/page";
-import OccasionTable from "../OccasionTable";
+import OccasionTable from "./OccasionTable";
 
 type DialogGeneralTabProps = {
   course: Course;
