+++ components/CourseModal/DialogTabs.tsx
