+++ components/CourseModal/DialogTimeChart.tsx
