+++ components/CourseModal/OccasionTable.tsx
